+++ scripts/tests.sh
@@ -105,7 +105,7 @@ EVM_VERSIONS="homestead byzantium"
 
 if [ -z "$CI" ]
 then
-    EVM_VERSIONS+=" constantinople petersburg istanbul berlin london paris"
+    EVM_VERSIONS+=" constantinople petersburg istanbul berlin london paris shanghai"
 fi
 
 # And then run the Solidity unit-tests in the matrix combination of optimizer / no optimizer