+++ test/libsolidity/syntaxTests/natspec/docstring_struct.sol
@@ -1,0 +1,12 @@
+contract C {
+    /// @title example of title
+    /// @author example of author
+    /// @notice example of notice
+    /// @dev example of dev
+    struct Example {
+        string text;
+        bool valid;
+        uint256 value;
+    }
+}
+// ----