+++ test/tools/yulInterpreter/EVMInstructionInterpreter.cpp
@@ -381,6 +381,7 @@ u256 EVMInstructionInterpreter::eval(
 	case Instruction::JUMP:
 	case Instruction::JUMPI:
 	case Instruction::JUMPDEST:
+	case Instruction::PUSH0:
 	case Instruction::PUSH1:
 	case Instruction::PUSH2:
 	case Instruction::PUSH3: